+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Full documentation for rocLibrary is available at [https://rocm.docs.amd.com/pro
 * Extended support to rocAL's video decoder to use rocDecode hardware decoder
 * Setup - installs rocdecode dev packages for Ubuntu, RedHat, and SLES
 * Setup - installs turbojpeg dev package for Ubuntu and Redhat
+* rocAL's image decoder has been extended to support the rocJPEG hardware decoder
 
 ### Resolved issues
 * `TurboJPEG` no longer needs to be installed manually. It is now installed by the package installer.